Fix linting and code format
